requirements.txt
PyMuPDF==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
Pillow==10.1.0
configparser==5.3.0
tqdm==4.66.1